BlogService/requirements.txt
absl-py==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dm-tree==0.1.9
einops==0.8.1
etils==1.12.2
fastapi==0.115.12
fsspec==2025.3.2
googleapis-common-protos==1.63.1
greenlet==3.2.1
h11==0.16.0
idna==3.10
importlib_resources==6.5.2
kazoo==2.10.0
numpy==2.2.5
promise==2.3
protobuf==3.19.6
psutil==7.0.0
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tensorflow-datasets==4.8.3
tensorflow-metadata==1.13.0
termcolor==3.1.0
toml==0.10.2
tqdm==4.67.1
typeguard==2.13.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
zipp==3.21.0
